WIP21: last step toward full POM flattening
It's overkill, because we only need to do it in order to know about
dependencies. So from this point forward, I'm going to narrow the scope
of the Resolver contract only to glean the dependency list, and no
longer try to do a full-blown POM model computation like Maven does.